day 7: test ok, input ko

+++ day_07/data.h
@@ -10,8 +10,8 @@
 #define MAX_NUMBERS 200
 
 typedef struct {
-    int total;
-    int* numbers;
+    long long total;
+    long long* numbers;
     int count;
 } calibration;
